@@ -1,11 +1,14 @@
 ﻿using System.Collections.Generic;
 using System.Net.Http;
+using System.Text.Json;
 using System.Threading.Tasks;
 
 namespace SpotifyApiExplorer.Interface
 {
     public interface IApiRequestService
     {
-        Task<string> PostForm(HttpClient httpClient, string url, List<KeyValuePair<string, string>> nameValueList);
+        Task<string> PostFormAsync(HttpClient httpClient, string url, List<KeyValuePair<string, string>> nameValueList);
+
+        Task<JsonElement?> GetAsync(HttpClient httpClient, string url);
     }
 }
